requirements.txt
requests
pyYAML
docopt
